+++ src/index.ts
@@ -1,1 +1,4 @@
 export * from './client';
+export * from './http';
+export * from './zk';
+export * from './api';