+++ typing/zkbasCrypto.web.d.ts
@@ -2,6 +2,8 @@ declare module '@bnb-chain/zkbas-js-sdk/zkCrypto/web' {
   export function ZkCrypto(): Promise<Z>;
 
   export interface Z {
+    cleanPackedAmount: (amount: string) => string;
+    cleanPackedFee: (amount: string) => string;
     getAccountNameHash: (accountName: string) => string;
     getEddsaPublicKey: (seed: string) => string;
     generateEddsaKey: (seed: string) => string;