+++ docs/CHANGELOG.md
@@ -1,3 +1,96 @@
+# [2.0.0](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.3.0...v2.0.0) (2022-09-19)
+
+
+* refactor!: New API ([fa2076c](https://github.com/bnb-chain/zkbas-js-sdk/commit/fa2076cf72201ffb4dee51fd5b9d31dec9f169a4))
+
+
+### Features
+
+* add getAccountNameHash methods ([#14](https://github.com/bnb-chain/zkbas-js-sdk/issues/14)) ([1db8eef](https://github.com/bnb-chain/zkbas-js-sdk/commit/1db8eefbb4a7f021be93d05d75109eae4caab287))
+
+
+### BREAKING CHANGES
+
+* Modify many API
+
+# [1.3.0](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.2.1...v1.3.0) (2022-09-19)
+
+
+### Features
+
+* Remove Price API And add assets API ([76d18f3](https://github.com/bnb-chain/zkbas-js-sdk/commit/76d18f302c3b1eb75a2743e0fa6f782b251472fd))
+
+## [1.2.1](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.2.0...v1.2.1) (2022-09-15)
+
+
+### Bug Fixes
+
+* SendRawTx return type ([1f06016](https://github.com/bnb-chain/zkbas-js-sdk/commit/1f06016fbdf7518715f03ea49ab447ab567d0aef))
+
+# [1.2.0](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.1.6...v1.2.0) (2022-09-13)
+
+
+### Features
+
+* Axios response type ([26961f8](https://github.com/bnb-chain/zkbas-js-sdk/commit/26961f80a72f4f2041608517a7d1b3c7f8dcd549))
+
+## [1.1.6](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.1.5...v1.1.6) (2022-09-13)
+
+
+### Bug Fixes
+
+* TxType export ([75bb0a4](https://github.com/bnb-chain/zkbas-js-sdk/commit/75bb0a41bc5faac3d7e812c2e1d2e62e993f25ff))
+
+## [1.1.5](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.1.4...v1.1.5) (2022-09-13)
+
+
+### Bug Fixes
+
+* Remove Comment ([e3f335c](https://github.com/bnb-chain/zkbas-js-sdk/commit/e3f335c018a458d8a16b4996e2418c61f4a81318))
+
+## [1.1.4](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.1.3...v1.1.4) (2022-09-09)
+
+
+### Bug Fixes
+
+* Post and Get params type ([54537b9](https://github.com/bnb-chain/zkbas-js-sdk/commit/54537b9b86452a7eda65bfb7e393d86e5cafeca3))
+
+## [1.1.3](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.1.2...v1.1.3) (2022-09-09)
+
+
+### Bug Fixes
+
+* Export constants ([dbae7e0](https://github.com/bnb-chain/zkbas-js-sdk/commit/dbae7e0e105e12eb6c7255386fe41770dbd60888))
+
+## [1.1.2](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.1.1...v1.1.2) (2022-09-09)
+
+
+### Bug Fixes
+
+* Export d.ts ([547a2ef](https://github.com/bnb-chain/zkbas-js-sdk/commit/547a2efd6e8b83a872949996624571431c361f3f))
+
+## [1.1.1](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.1.0...v1.1.1) (2022-09-08)
+
+
+### Bug Fixes
+
+* CurrentPrice API request param ([fee006e](https://github.com/bnb-chain/zkbas-js-sdk/commit/fee006ea280603e2747e6d9fb2358eda3c8c67fe))
+
+# [1.1.0](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.0.1...v1.1.0) (2022-09-08)
+
+
+### Features
+
+* CurrentHeight API ([d56a08b](https://github.com/bnb-chain/zkbas-js-sdk/commit/d56a08bc534daa528eb16288e3cc3434ca5ee34a))
+* New wasm method ([e200d89](https://github.com/bnb-chain/zkbas-js-sdk/commit/e200d89a12259dac00ee93fa2809f52d29f21235))
+
+## [1.0.1](https://github.com/bnb-chain/zkbas-js-sdk/compare/v1.0.0...v1.0.1) (2022-09-05)
+
+
+### Bug Fixes
+
+* Creatred time type ([44feb1f](https://github.com/bnb-chain/zkbas-js-sdk/commit/44feb1f3a825d8fc9075fb55f27b0dac0b62c798))
+
 # [1.0.0](https://github.com/bnb-chain/zkbas-js-sdk/compare/v0.11.0...v1.0.0) (2022-08-30)